requirements.txt
# RA-OCR Dependencies

# Environment & CLI
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0

# PDF Processing
pymupdf>=1.23.0

# HTML Rendering
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM APIs
google-generativeai>=0.8.0
openai>=1.0.0

# Image Processing
Pillow>=10.0.0
